All telemetry now includes OS type and ASA version
